+++ Staging/AudioDxe/AudioDxe.h
@@ -55,7 +55,7 @@
 #include <Protocol/HdaControllerInfo.h>
 
 // Driver version
-#define AUDIODXE_VERSION        0xC
+#define AUDIODXE_VERSION        0xD
 #define AUDIODXE_PKG_VERSION    1
 
 // Driver Bindings.